Alternative instance for Parser
All of the pieces for this were already there, so I upgraded the Monad
instance to a Monad with Alternative instance. I'm pretty confident that
this is a lawful instance (at least to the extent that the existing
SemigroupK instance was), but since it's dodgy to define equivalence
for parsers, it looks like it won't be straightforward to try to check
laws.

This is _not_ a binary compatible change. The ParserSimigroupK instance
was removed, as it would now clash with Alternative instance.