Improve horizontal whitespace
The character class includes various chars.

The previous test was unlikely to fail, as it checked only
arbitrary chars, most of which are obviously not whitespace
even to the most broken implementation.

Note that it's not helpful to display an arbitrary
horizontal whitespace char after a failed test, though
I imagine it was a minimized input.